iot_web_app/backend/requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
pyserial==3.5